1. 优化 JobHandlerInvoker 的实现，增加重试机制 2. 去除作业异常，自动暂时作业的机制。
