+++ src/main/java/cn/iocoder/dashboard/framework/quartz/core/enums/JobDataKeyEnum.java
@@ -7,6 +7,8 @@ public enum JobDataKeyEnum {
 
     JOB_ID,
     JOB_HANDLER_NAME,
-    JOB_HANDLER_PARAM
+    JOB_HANDLER_PARAM,
+    JOB_RETRY_COUNT, // 最大重试次数
+    JOB_RETRY_INTERVAL, // 每次重试间隔
 
 }
