+++ src/main/java/cn/iocoder/dashboard/modules/infra/controller/job/vo/job/InfJobBaseVO.java
@@ -23,6 +23,14 @@ public class InfJobBaseVO {
     @NotNull(message = "CRON 表达式不能为空")
     private String cronExpression;
 
+    @ApiModelProperty(value = "重试次数", required = true)
+    @NotNull(message = "重试次数不能为空")
+    private Integer retryCount;
+
+    @ApiModelProperty(value = "重试间隔", required = true)
+    @NotNull(message = "重试间隔不能为空")
+    private Integer retryInterval;
+    
     @ApiModelProperty(value = "监控超时时间", example = "1000")
     private Integer monitorTimeout;
